Move MetadataDebugStringOptions to Infastructure and make DebugView public
Fixes #20381
